add : 게임 매니저 커스텀 에디터 테스트

+++ Assets/Scripts/Misc/Data/DataCloud.cs
@@ -43,7 +43,9 @@ public static class DataCloud
 
     public static void DeletePlayerData()
     {
-        PlayerPrefs.DeleteKey("PlayerData");
+        if(PlayerPrefs.HasKey("PlayerData"))
+            PlayerPrefs.DeleteKey("PlayerData");
+
         playerData = null;
     }
 }
